Don't remove syncs upon logout
MEGAsync keeps local sync configs even if the user logs out. The goal is
to allow to restore syncs upon a new login (in the same account).

The SDK should not remove (`sr`) the heartbeat record in that case.
Otherwise, when the sync/backup is restored and re-enabled, the `sp`
command will fail to update the state (so it remains as failed due to
logout).

Note that any sync activity will send a heartbeat (`spbh`) and the API
will revive the backup, but it will keep the previous state (which is
outdated).