checkCloudPathForMovesRenames, also check for a completed moveContentTo
previously we tried to avoid doubled up logic by letting the final resolve_rowMatched finish the moves.  But, we won't get there if the fsNode moved again in the meantime.  This code just checks the cloud side completed and sets up to recognize the next move in a chain of moves.

Remembered the handle moved so we can use it to confirm the end of the move
findLocalNodeByFsid no longer needed - we were checking synced and scanned, but then quitting the function if it didn't match on synced_fsid.  So, we only need to search the synced_fsids.   That means this function is redundant, removed (well, commented for now)