SDK-2398. Minor changes from sync rework (24th stage) - bring SRW sync states and transitions
SRW has 6 states for sync, whereas the old system only had two.
This merge brings the interfaces for working with the 6 states, but the sync code is unchanged, and still only works with two
This allows MEGAsync to be updated to use the new (and simpler) interfaces.   The old ones had gotten a bit complicated over time.

Apps are always informed of sync addition by sync_added callback. That goes for auto-loaded syncs/backups or by request.  The intermediate layer function to add a sync just reports success/failure, but not the sync addition itself.
Functions to specifically enable/disable syncs, and the callbacks specifically for informing of enablement/disablement, are removed.
onSyncStateChanged() delivers those changes instead.  setSyncRunState() is for requesting state changes.

removeSync() is just one function now, no overrides

removeSyncs() is itself removed, as it was unused, and that avoids awkwardness with backupDestination, and which error it might deliver if there were multiple.
There is no longer an 'active' flag for syncs.  They are either Running or Disabled.  (the other 4 states cannot occur yet)

SyncAdded enum is removed.  The app only needs to know it's added, and what its state is.
Added SyncRunningState: the new six possible states for a sync

removed:  isEnabled(), isActive(), isTemporaryDisabled()
added: getRunState()

sync_auto_resume_result() removed, as sync_added() covers that now

added syncupdate_syncing() which indicates if there is syncing activity going on.

Internally syncs are still using syncstate_t, but it's no longer supplied to the app or even to the intermediate layer.  When SRW is merged, that will be gone.  mRunState is updated in parallel with it for now.

Various other small related things brought from sync rework, to help the diffs line up between the branches.

Removed resetSyncConfigStore() as it was only called from one place, made that code inline (and match SRW)

Adjusted resumeResumableSyncsOnStartup and related functions to be like SRW.  Except that, syncs must be loaded and working before actionpackets start here.   In SRW, that happens after statecurrent.

Tests are updated to use the updated interfaces.