Fixes for test MovedSyncedFileWhileDownloadInProgress after other sync changes
Proper tracking of which nodes are involved in moves, so we can always find the moved-from node, from the moved-to node.  This is not always doable via fsid, for complicated cases where the local file was moved again etc.

Downloads to know which fileFingerprint is ok to overwrite (the file that was previously synced at that node)

Detect and execute (move) the completed downloads in syncItem() before splitting into the eight CSF cases  (previously, only checked in resolve_downsync(), but we don't always go down that path).  New function Sync::syncItem_checkDownloadCompletion for that.

Perform the moveContentTo() and other data structure maintenance from checkForCompletedCloudMoveToHere() (previously only during SRT_CSF case)