SDK-2403. Add new separate function for moving or removing old backup Nodes from the Vault
Added moveOrRemoveDeconfiguredBackupNodes. This operation needs to be available anyway for a full-featured SDK
removeSync is modified to just remove the sync/backup configs.
Added OnFinishOneShot class to enable Qt apps to easily pass a temporary listener for a Request, and take action on the onFinished callback on the GUI thread.
Apps can use OnFinishOneShot to easily call moveOrRemoveDeconfiguredBackupNodes for backups, should removeSync() succeed.
Now the apps have the option to leave Backup nodes in place intentionally.
removeSyncs() function is itself removed, it was not used.