Fix problems in NoD code, spotted while trying to integrate with SRW
pointless extra overload of FileFingerprint::serialize, that dangerously mixes overloads and overrides
serialization issue that breaks compatibility - size_t is not a fixed size itself.
attrstring and nodekeydata can be longer than 16k.  Again mistakes with serialization
random reloads of the Node tree are not a good idea.  That sort of thing cannot be decided in an extremely low level function