Avoid processing filesystem notifications from below ignored folders
This lets us get to the fully-scanned state fairly quickly for an existing sync that is a parent of our own database folder, for example.
Also, for the purposes of sync heartbeats, only consider it "scanning" until it first completes scanning.  Later scanning (even though we do it) interferes with the % display in BC