SDK-2501. Syncs to detect when the local filesystem takes liberties with file names, changing them unexpectedly
eg. Synology, connected as a networked SMB drive on windows, will report success renaming a file to have trailing spaces, and also silently remove those trailing spaces.

For this case, report a stall so the user can decide to rename the file to be Synology-compatible, or shift the sync elsewhere.