Ensure that mAuthRingsTemp is properly purged after initialization
Depending on the state of the key cache and the authrings, in many cases
mAuthRingsTemp could be active after initialization, not allowing to
properly update keyrings nor the new ^!keys attribute.

The easiest ways to reproduce this problem (before this commit) were:
- Resuming a session with all keys cached
In that case, not getua nor getpubkey would send a request to the
server. Since mAuthRingsTemp is only purged in the response of commands
sent to the server, all nor authrings were kept active.

- Login+fetchnodes with the signature of some keys already verified
In that case, getua and getpubkey would send requests to the server.
However, if the signature was already verified in the keyring, it was
not verified again and since mAuthRingsTemp for signed keys is only
purged in trackSignature(), those authrings were kept active.

This commits ensures that mAuthRingsTemp is not kept active if no
requests are sent to the server. The check of the signature is forced if
mAuthRingsTemp is active (to be able to purge it). For not signed keys,
if mAuthRingsTemp is active, it's checked if it has to be purged even if
the key was already being tracked.