SDK-2528. Bring back Node::unserialize
This function was moved from node.cpp to megaclient.cpp and renamed as NodeManager::unserializeNode. However

almost all the code in it is specific to Node, not NodeManager
it's 150 lines that should exist in node.cpp, not megaclient.cpp
megaclient.cpp is already huge, no need to add code that shouldn't be there
Node should have symmetric serialize/unserialize functions, same as all the other Cacheable objects. Bring back this nice pattern.
Restore the git history so we can compare the contents of Node::unserialize with older versions of it.