Fix for cases involved when a .megaignore file is edited, and then the upload is queued but before the upload starts, the file is removed from disk.
Ignore move logic when a .megaignore file's SyncNode should be deleted.  We can't detect such moves due to the uncertainty of what nodes are excluded or included, and hence should be visited.

When sync transfers fail, inform the sync thread to scan the folder again (for uploads) or sync it again (for downloads) so that we don't immediately set up the exact same transfer again.

Track whether fopen() succeeded for uploads, so that we can more accurately report why they fail (we were erroneously saying the size/mtime changed, when in fact the file was missing)

Log some info on why uploads may need to be replaced, in case we need to diagnose those again

Also converted a related pointer to a smart pointer
Also added a facility to easily log fingerprint data