Create *!keyring even with support for ^!keys
Segmentation fault detected in the client that confirms a new account
with support for ^!keys -> the `upv` doesn't set the legacy attribute
and, in consequence, the signing/chat key objects are not initialized.
As soon as this keys are used, client crashes.

This commit restores the initialization of that attribute, which
improves the compatibility with legacy clients, which expect that
attribute to be present.