Assume that authrings are empty during the initialization
During account creation, loadAuthrings() tries to get legacy attributes
(because ^!keys is still not created). When the response arrives,
initializekeys() has been already executed, so generation is 1 and the
-9 received in getua doesn't initialize the authrings and we get a
failed assertion in fetchContactsKeys().

Alternatively we could initialize generation to 0 to let the response of
legacy attributes initialize the authrings. In that case, the first
generation sent to the server would be 1, instead of 2, but on the other
hand that change could have other side effects (for example it could
allow to call upgradeSecurity multiple times).