SDK-2553. Introduce treatAsIfFileDataEqual() function for deciding if it's ok to clone a cloud Node or skip an upload/download
Bringing this change from Sync Rework to develop.

This function to only allow that optimization for suitable media files.  This prevents erroneous elides of uploads/downloads where just a small difference exists between eg. text/code files, where the difference is in the gaps between fingerprinted areas.

Use treatAsIfFileDataEqual to decide if we can put multiple Files in a single Transfer (hence the multimaps)
Use treatAsIfFileDataEqual to decide if manual impl downloads can be skipped because the file is already present
Use treatAsIfFileDataEqual to decide if syncs can skip uploads and clone nodes instead

Make the maps of transfers (by fingerprint) into multimaps (still by fingerprint)

Waiter to be shared_ptr<Waiter> so that we can pass it to objects that might outlive the source (eg on another thread)

megacli: bring an update to exec_du so we can see the breakdown by contents of a folder

Also use a smart pointer for asyncopencontext, it was a tiny change in the same area