SDK-2571. Delete the call to reload() from MegaClient::readnodes
This is the only non-NOD use of reload() and it is not helping. The conditions for calling that make no sense, and may be causing MEGAsync to reload the account ... not immediately but hours or days later, probably unnecessarily, and unexpectedly to the user, causing lots of complaints. It is also an obstacle to getting rid of the isCrashed flag and associated problems in MEGAsync. Deletion of this nonsensical section is one step towards solving those other problems.