Check share keys to prevent broken keys in Node::nodekeydata
If an invalid / outdated share key is used to decrypt a node and
nodekeydata is updated with the resulting decrypted node key, the node
can't be recovered at runtime, even if an updated share key is received
later (because the undecrypted node key is no longer available).
This MR prevent changes in nodekeydata if we are able to detect that the
share key is invalid (unable to decrypt the attributes of the node).