SDK-2615. Allow a sync to auto-restart if it was previously suspended because the local path was not available
This helps with
   - syncs that could not start on app start because the machine is starting up, and networked or other drives are not yet mounted but are part of the startup sequence
   - people that want to be able to remove and restore USB devices etc
   - networked drives that glitch or are offline for a while

Disabled drives will not be auto-restarted, as we have lost the sync state for those

We now store the fsid of the root folder, so we can double check it is unchanged.  And so that we can determine the sync database name even if the folder is not present and openable currently

A new error code is added if the fsid of the root folder is not the same as before

With this change in place, a sync on a USB is supended when the USB is removed, and auto starts again when the USB is re-inserted.