Adjust condition to add share keys to KeyManager
Trusted keys should never change, but if at some point that policy
changes, or a secure client replaces the key by mistake, avoiding to
replace the key could cause problems, because probably the correct
approach would be to accept the key (if it has been received from the
secure channel).

The replacement of keys coming from the legacy channel (action packets),
which is the most problematic, is managed in the specific code path at
`mergenewshare()`, and in other cases it's already checked if the key
exists before trying to add it. So it could cause undesired side effects
if it's triggered at some point: ie. the key could be discarded here but
used in nodes or shares, since the return value is not being checked.

Better to allow the replacement at this stage, and monitor a new event.