Share folder with secure=false, originally shared by secure=true
If a "ok" has been created with secure=true (dummy/invalid, all zeros),
clients with secure=false that attempt to create a share on the same
node will send a different "ok" (the actual one) and the API will reply
with the all-zeros "ok". It will break the share key at the client with
secure=false.
This commit aims to avoid that corner case by sending "ok" with zeros if
the key is trusted/created with secure=true.