Find unverified outshares by email AND userhandle
Pending outshares (with non-contacts) are added to the
`keys.pendingoutshares` with email, but once the sharee becomes a
contact, we may search by userhandle. This commit ensures we search by
both, email and userhandle, since both can be used for
`keys.pendingoutshares` (handle is used when they are contacts).

Also adjusts:
- tests to new behavior of Share getters
- inaccurate documentation in MegaApi for getOutShares()