Cache the JSON we prepare for a Request (ie, batch of Commands) in case we need to resend it
Caching it guarantees that we resend the exact same JSON array, which is required in order for the API to support idempotence of these requests, in case network issues mean we lose the response, but the API did execute the request.
Also calculate and cache the idempotenceId (also called reqId) associated with that string at the same time, so they are managed together and always match.

move reqid from MegaClient to RequestDispatcher
move resetid from MegaClient to Utils

also some collected small sync code improvements