Fix DirectRead not being removed after processAnyOutputPieces() returns false
When processAnyOutputPieces() returns false, continueDirectRead is false
too. That value is given by pread_data returning value. If the
pread_data returning value is false, the transfer is being finished, and
the DirectRead node must be explicitly deleted.

There is two calls to processAnyOutputPieces(). One of them is correctly
deleting the DirectRead object, but there was a second call which
didn't. This commit fixes that situation.