SDK-2717. The idea is to keep the actual case statement code in the same place, same indentation, in the file so that future diffs looking back on this can still find the original author.
The case statement will be converted one by one, from the bottom of the loop

So, the tail of the loop (after the switch closes) is rearranged to be at the top, for minimal changes as those are converted.

One single function is converted as an example.

Benefits:

the operation's code to construct the Request object, and the code implementing it, are in one spot, making understanding, maintenance and development easier.
we will no longer have a 4k line switch statement which new people find hard to navigate
Diffs to look up issues will hardly be affected, as the feature-implementing code is untouched.