SDK-2743. Add downloadFileHandle as a Transfer member, so that on transfer resume, we can double check we are really still downloading the same file (and not a possible mismatch that just matches on FileFingerprint)
For downloads, when checking if we can resume a cached transfer, only allow it if the source file node handle is a match.
Add downloadFileHandle to the serialization/unserialization of Transfers
Update Transfer::unserialize to use CacheableReader.  (Transfer::serialize is left for another day so we can test this new unserialize against the old serialize)
FileFingerprint::unserialize() updated to be compatible with CacheableReader
Remove OLDUPLOADTOKENLEN which is is from 2016.  7 years later, there should not be any cached transfer db records using that
Added a few serialize/unserialize functions in CacheableReader to support data types needed by Transfer that were not there yet