Replace public Set preview solution approach to behave as FileLink
Previous approach was a behavior similar to FolderLink. That approach
has proved unnecesary and a lighter version such as FileLink is enough

The new approach doens't require an extra MegaApi instance, nor is
bound to TYPE_LOGIN requests.

In addition to the simplified approach, MegaApi interfaces and model
is also simplified in a way that is no longer required to start the
Public Set preview mode. With the new approach, any client application
may request directly a Public Set fetch and the mode will be set as a
inner process step.

On top of that, some relevant remarks:
- fetching Public Set which is already fetched and available, does not
     perform a request to the API and provides the already fetched
     Set. This implies that even if the Public Set stopped being
     Public, it will still be provided until a different Public Set is
     fetched, or the Public Set Preview is stopped.

- and exporting/disabling exports on Sets operations implement now a
     shortcuts to avoid requests to API on Sets which already present
     target status

- errors fixed regarding Set's ts update during Set export enable and
     Set export disable. In particular to the delayed ts information
     retrieved only via APs

- error fixed on SetElement copy and assignment ctors and functions

- integration tests as well as megacli adjusted to the new approach
     and interface

- new cases in the integration test to cover all new found errors
     regarding ts, order (i.e. SetElement), and APs updates