First attempt at making NOD thread safe so sync can access it directly from the sync thread.
Tried to use std::mutex but had to use std::recursive_mutex for now, due to NOD calling MegaClient which then calls back into NOD.

Basically, separated NodeManager into a public layer which locks the mutex, and the rest of it is the internal layer, which can freely call other internal functions.

Removed one NodeManager function which was basically a duplicate and had a misleading signature

Moved the nodeByHandle() undef check into NodeManager, from MegaClient (where it will be useful for sync).