Fix the passing of parameters in callbacks from the SDK
The default typemap in SWIG seems to be wrong. It tries to use the
pointer of the pointer that we pass as parameter and that causes a
crash. With this change, we use the same code but force SWIG to directly
use the pointer that we pass ($1) instead of passing its address to
SWIG_SetPointerZval.