Removed obsolete variables / functions
The functions getCurrentRequest/getCurrentError/... and the associated
variables activeRequest/activeError/... were there just to workaround
some problems in PHP bindings. They have been solved already so this
hack is no longer needed. It's extremely unlikely that any other code is
using these functions because they only worked during callbacks, to
provide data that is already available in the parameters of those
callbacks, so it doesn't make sense to use them in any app that can
directly use the parameters of the callbacks.