SDK-2783. Update MegaApi file permissions along with MegaClient and MegaFolderDownloadController
The permissions specified by the user via
MegaApi::setDefaultFilePermissions() are not honored by the SDK when it
downloads files/folders.

This is due to MegaClient::fsaccess no longer being a copy of
MegaApi::fsAccess.

Besides, MegaFolderDownloadController (as well as
MegaFolderUploadController) has an independent FileAccess object.

This commit fixes it, so MegaApi also updates MegaClient:fsaccess
object, and MegaFolderDownloadController sets the default file and
folder permissions from MegaApi.