SDK-2768. User Journey Tracking Support
- Fetch, store, and populate url with the journeyid
- Modify log function to allow a boolean 'journeyTrack', and if true
then submit the journeyId with the `log` API request.
- Modify log function to accept an optional viewID.
- Attach the current millisecond to the log api request if journeyTrack
option is set.
- Expose method for apps to generate viewIds.
- Add methods to cache JourneyID value and tracking flag on-off
- Load JourneyID and tracking flag from cache. JourneyID value is always
reused (until a full logout). Further JourneyID values parsed from "ug"
and "gmf" command results are only used to set tracking flag on (if it
was set to OFF and there is a valid JourneyID value on the result) or
off (if it was set to ON and there is no JourneyID value on the result).