Make sync cloud node creation work the same way as Moves, which has worked out well
There was a problem when some unlucky timing could cause a locally user created folder, and then user renamed, could end up with folders of both names remotely and locally.
On cloud folder creation, we need to set up the syncedHandle and syncedFsid so that if the local item was moved/renamed even before we see the created cloud node,
we still have enough info in the data structure to deduce what happened and perform the required move.