Skip fetching keys step and directly initialize keys
Keys are now loaded from ug response so they could be initialized
directly after the ug response is processed.
Removed code to wait for the last signature to start initializing keys.
Moved event sent in case of a failure puting keys from
CommandPutMultipleUAVer to a lambda when calling putua.
A piece of code setting the received RSA pub key in CommandPubKeyRequest
while getting keys has been removed, it was not used and it is retrieved
in the ug response if it exists.
Moved some sanity cleans regarding the keyring and user from fetchkeys
(now removed) to initializekeys.