SDK-2813. Adjust error logging on file open, not all failures to open are really errors
Particularly when syncing, the logs are now full of error messages that
are not actual problems. Often for these cases the code is just checking
for the presence or absence of a file, and it will take action based on
the results.

Adjust the relevant functions to take a parameter so that the caller can
say whether to log or not, or supply the conditions under which to log
(some of them are supposed to log any error except file not found)

Supply said flag from all relevant call sites, as appropriate for that
situation