SDK-2820. Fix CREATE_ACCOUNT flow so it cannot be broken by other requests.
When a CREATE_ACCOUNT is requested, the flow is complex:

1) First, the account is created.
2) When the response is received, and processing an AP, the command
CommandSendSignupLink2 is executed.
3) Finally, the "sendsignuplink" command is processed. The
client->restag is set to the cmd->tag (the tag associated to the
CommandSendSignupLink2, defined upon creation from the client->reqtag
value).

If the client->reqtag was different to the client->restag on step 2, the
cmd->tag for CommandSendSignupLink2 would be different than the one
where the initial request (CREATE_ACCOUNT) was stored on the RequestMap.

This can happen if there is another request in the middle, between steps
1 and 2. For example, if there is a recurrent RetryPendingConnections
from megaCmd, like this case.

This commit fixes that problem by passing the client->restag, where the
CREATE_ACCOUNT request was mapped, to the CommandSendSignupLink2
command, so the cmd->tag and the client->restag are the same, and the
sendsignuplink_result function can retrieve the original request from
the RequestMap.