SDK-2900. Remove the Anomalous name reporting
Although this was implemented in the SDK, it was never used in any app. And it has become much less relevant now that LocalPath and SRW can deal with these paths easily and correctly on windows.

Let's remove this code so we don't have to continue maintaining it here and on SRW.   If it proves necessary in future, we can bring it back from git.