SDK-2918. On SRW download completion, update the row synced fields so that we can chain moves that occurred during the download
We convert it to the same system as is working now for uploads/moves etc, which is to set the row as synced with the results of the transfer, even if those nodes/FSnodes have moved since, so that we can correctly detect those subsequent (but in-parallel) moves and perform them also afterwards. The old system of only detecting the row as synced once all the nodes are lining up doesn't work if they don't get to an aligned situation, due to the user moving things before we hit that state.