SDK-2928. Fix calculation of last chunk size
When there is a last chunk of small size, the current calculated chunk
is expanded to cover that small size remaining to complete the transfer.

Before a recent change, the value of the nextChunkSize in
RaidBufferManager::nextNPosForConnection() could be used to check if the
nextChunk is the lastChunk. After that change (which fixed that logic),
nextChunkSize couldn't be used as the lastChunkSize anymore.

This was causing a bug by updating the npos to the maxpos for all the
next chunks, when they were not the last chunk, and getting a very large
next chunk (if maxpos = 1GB, then the next chunk was 1GB size).

This commit fixes the issue by calculating a lastChunkSize (maxpos -
npos) and use it for the purpose of expanding the last chunk if this is
too small.