SDK-2943.  SRW: Moving a file out of a sync, and then immediately back in while the sync is still processing the first change
For both cases, we need to set up the synced-fsid and synced-nodehandle for the appropriate row at the earliest opportunity, because if the user makes their change fast enough then we don't get a chance to see the row cleanly matched up with all 3 nodes, and we don't mark it synced.  Then when the 2nd change arrives, our data structure doesn't contain the updated info to make a correct decision

Also including a small update for QTMegaRequestListener, so it can be used for MegaApi calls that the user needs to get a result back from the Request's member functions, after the operation completes.  This was needed in some recent dev investigations and will be generally useful