Round minSpeedPerConnection to RAIDSECTOR and return false on pread_data if transfer is finished
-> minSpeedPerConnection needs to be padded to RAIDSECTOR if the
transfer is raided.

-> pread_data returned TRUE if the DirectRead should continue, but FALSE
only when it failed because of being incomplete. If it ended OK, it
returned TRUE so the DirectReadSlot could be deleted by itself when all
the requests are set to REQ_DONE state, as the calculated position is
greater or equal than the final position.

However, that's not really necessary and has potential issues. It has
been changed.

-> Added extra logging information and removed a couple of unneeded
instructions.