Delete DirectRead with a deleted transfer on DirectReadNode::retry()
-> If a DirectRead has a deleted transfer, it must be deleted and
removed from the DirectReadNode::dr_list.

-> It can't be possible that the DirectReadNode::retry, called from a
given DirectReadSlot, with a DirectRead which has a deleted transfer,
had been called from the same DirectReadSlot belonging to that broken
DirectRead. Prevention has been added on DirectReadSlot to avoid such
situation.

-> Hence, the affected DirectRead with a deleted transfer can be deleted
on iteration from DirectReadNode::retry. That deletion will remove it
from the list. It's the same mechanism as the one on
MegaClient::abortreads.