Fix recalculated npos to be divisible by RAIDSECTOR
On raid.cpp::nextNPosForConnection() for RAID transfers, when lastChunk
< MIN_LAST_CHUNK, but currentChunk + lastChunk > MAX_LAST_CHUNK, npos is
set to half of the remaning size (currentChunk + lastChunk).

This means that there will be two last chunks of similar size.

However, the first one of the chunk last chunks, currentChunk, had set
the npos with a size not divisible by RAIDSECTOR. This made some
transfers not to end correctly. This commit fixes that issue.