Reduce the threshold in SdkTestCloudraidTransferResume
When transfers are interrupted like in that test, partially downloaded
chunks are not cached. So the cached progress can be quite lower than
the progress reported by the transfer.

In most cases, the cached progress will be higher than the threshold set
in the test, but not always. Reducing that threshold should prevent
false alarms.

Even with the reduced threshold, the test should catch genuine failures
resuming transfers because the check is fast enough to get the progress
of the resumed transfer before it has enough time to download new data
(at least during my testing), so I think that it's extremely unlikely
that the test succeeds because the resumed transfer get a temporary URL
and downloads more than 2MB during the 100ms that it waits before
getting that information.