Avoid race updating email
If the app calls `MegaApi::getMyEmail()` in the callback
`onRequestFinish()` of the request `MegaApi::changeEmail()`, the email
has been already changed internally, but the intermediate layer have not
received the `MegaApp::users_updated()` that will update the cached
email.

To avoid that race condition, the SDK needs to update the email
immediately. This commit ensures to update in MegaApi when it's updated
in MegaClient thanks to `reportLoggedInChanges()`.