Update cached email before completion of fetchnodes
When the client starts a new session and `f` command returns a cached
response, the subsequent action packets don't produce notification for
the app. In result, the cached email in the MegaApi is not up to date
upon `onRequestFinish()` for fetchnodes.

This commit solve the issue by explicitely call
`reportLoggedInChanges()` when the action packet corresponding to the
email is set.