Fix current logging information and add some extra info for TransferSlot
-> Change unmeaningful data (such as the current transfer position) to
more useful info (raid part progress, overall completed progress, etc)
-> Add connection indeces to TransferSlot logging instructions, and
progresscompleted when updated, not after a REQ_SUCCESS
-> Change FSLogging::logOnError to FSLogging::noLogging when trying to
open a file when the localname is different to the localfilename (this
happens always that we use temporary names for a transfer). This
"logOnError" will happen when the file doesn't exist (the destination
path doesn't exist yet), but that's the expected. If we kept this, we
would always have that error log line unneccesarily.
-> Clarify a few other logging instructions.