Fix TransferSlot::MAX_REQ_SIZE on Windows to be consistent with the macro definition
There is a macro definition for TransferSlot::MAX_REQ_SIZE depending on
the system.

TransferSlot::maxRequestSize object attribute is populated from that
static value.

However, on TransferSlot constructor, that value is redefined when we
can retrieve RAM information (global memory status).

If memory values are lower to some boundary, the
TransferSlot::maxRequestSize object attribute is decreased to some lower
value.

If that doesn't happen (the values aren't lower to that boundary) then
the maxRequestSize value should be set to TransferSlot::MAX_REQ_SIZE.
However, it is set to a hardcoded value which was the same as the
initial TransferSlot::MAX_REQ_SIZE, but it isn't the same anymore.

That definition (on the "else", i.e., we have RAM information, but the
memory value is below the lower boundary) should be populated from
TransferSlot::MAX_REQ_SIZE, and not from a hardcoded value.