SDK-3010. Occasional uploads that don't complete
Due to a race condition: the new cloud Node can arrive on the client thread, while the sync thread is processing that same folder. If we update the sync data structure with data from the completed upload (in case 0 (moves)), but we didn't have that Node available already when the folder triplets were calculated, then we also visited the node again for case 1 (non-move actions), and saw the XSF case, and created a 2nd upload.

The next pass of the tree would see the in-sync row though, and we may not even visit that folder again if all child items are considered in sync.

Easy to fix by considering that child item processed by case 0, so case 1 is not processed on that pass.