Add test deleting a request's listener before finishing the request
RequestTracker is used as the request listener for SdkTest requests:
MegaRequestListener type.

This new test start a simple request (importing a file link) and deletes
the listener before the request is finished.

Besides, MegaApi::removeRequestListener() is now called within
RequestTracker destructor.

Finally, when the request is finished and fireOnRequestFinish() is going
to be called, the listener obtained by request->getListener(), which has
been deleted, should be NULL now thanks to the removeRequestListener()
call, avoiding to dereference a corrupted pointer and ending the test
without a runtime error.

The changes on RequestTracker destructor are also useful to ensure that
the MegaApi::removeRequestListener() call on a MegaRequestListener
destructor doesn't affect normal executions.