Avoid converting a private chat into a public one.
Due to an API race condition we could receive an mcpc AP with m==1
after receiving a prior one with m==0, the AP's would be triggered in
API by different clients so this use case is legit.

We need to avoid this as the way as chat encrypts/decrypts msg's and
title depends on mode, so setting an invalid mode could ends up with
unencrypted/decrypted messages.