Add utility functions for fingerprint app prefix addition/removal
Added in this commit a safety check comparing results of the new
functions as part of an existing integration test.

The goal of the safety is to easy the reviewer a means of validating
and verifying the replication of the functionality. As such, this
safety test will be removed in a follow-up commit