SDK-3051. Convert share and link related members of Node to unique_ptr to clarify ownership
refactoring ahead of LRU:

With the LRU feature coming up, Node will be deleted at arbitrary times by the last surviving shared_ptr to it. Its destructor therefore should not be touching other data structures outside of Node. It did contain several raw pointers to share related things, and ownership was not clear - could ~Node() safely delete those?

This ticket is to convert those to unique_ptr to make ownership explicit, and make sure those pointers are not also copied to other data structures in the program. It also tidies up the destructor considerably, as those will be deleted automatically.