SDK-3066.  Deal with Backups that have old duplicate names in the cloud
Those are probably historic from the pre-SRW sync code, but real users reported they hit this case.
Since the stall items window in MEGAsync can't alter the Vault, we need to address that directly in the sync.
Fortunately, the cloud should be a direct exact copy of local so we can easily determine what to remove, and not be concerned about moves etc.