SDK-3078. Restore TwoWay_Highlevel_Symmetries as a single test that exercises all the combinations.
It was taking 45 minutes instead of 10 when split into 8 tests (and run the normal way, sequentially and only requiring 3 test accounts).  Also, it's easier to debug and maintain, not to mention explain, as a single test.

If it causes the parallel runs to to a bit longer, the parallel runs always have the option to queue it by itself for one executor.