SDK-3086. Remove mNodesWithMissingParent member of NodeManager
After discussion with NOD team, this member should not be a member of NodeManager class. Rather, its use is entirely local to the readnodes() function, and so should be a local variable there (and passed to NodeManager as an argument for the moments where it's relevant).

That keeps the design simpler, we don't need to worry about thread safety of that member, and anyone learning this code doesn't need to worry about its state over the lifetime of the program - rather, only during readnodes(). Considerably simpler.

Also noticed that the event relating to the map being non-empty is getting triggered, 2 million times per day since May.   Now we will get an assert() to alert us to when this is happening, so it can be investigated when we have the relevant evidence before us.