SDK-3097. Add .Spotlight-V100 to the list of do-no-sync files
That folder seems to not be listable (under some circumatances at least), and as it is automatically maintained by Mac, based on the contents of the external disk, not something that is suitable to sync anyway.  Downloads would break spotlight, and any change on the disk would result in this folder's content also changing and more uploads.  And, if left out, the mac will regenerate it anyway if needed.