Search for share keys in key manager
In some scenarios, the sharekey is not set in the Node, but it's
available in the KeyManager. This commit aims to look for the sharekey
in the KeyManager when the account is already migrated (secure=true and
generation>0).

Some examples:
1. versions inside outshares, received in response to `f` in the `t.f2`
element -> the sharekey in the `mNewKeyRepository` is discarded after
processing `t.f` (files/folders).
2. all nodes inside inshares that are received when the sharekey is
already available (pending shares, established before the contact
relationship) -> key is available, but not established for the root of
the inshare yet.