Remove raidproxy threads
-Raidproxy threads were meant to run some functionality in parallel.
Initially, those threads handled all the sockets (opening them, writing
on them and so on).

-As this can only be done on the MegaClient thread with the waiter
dedicated to that task, less functionality were parallelized on the SDK
implementation.

-Finally, taking into account that most of the raidproxy functionality
needed to be synchronized with TransferSlot::doio (all the functionality
affecting HttpReq objects: preparing them, managing the REQ_FAILURE
state and so on), it has been considered to remove the threads and
synchronize everything.

-The big part of the CR improvements are still intact, as they are based
on parallel connections, which are still present.